requirements.txt
pip >= 7
django >= 1.10.1, < 1.11
mysqlclient==1.3.10
sorl-thumbnail
django-resized

#django-mysql
#mysqlclient==1.3.10

#django >= 1.7, < 1.8
#PIL==1.1.7
#MySQL-python==1.2.3

#dj-database-url==0.3.0
#dj-static==0.0.6

#django-toolbelt==0.0.1
#gunicorn==19.3.0

Pillow
django-dbbackup
#psycopg2==2.5.4
#pyDatalog==0.15.2
#pyfacebook===1.0a2.1
#spyder==2.0.3
#static3==0.5.1
#virtualenv==12.0.7